# docs/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Virtual Shelf Book Service API
  version: 0.10.0
  description: |
    Virtual Shelf MSA의 Book Service 계약이다.
    알라딘 API 기반 외부 도서 검색, 개인 서재 관리, 독서 진도, 문장 스크랩, 동물 사서를 제공한다.

    표지 OCR과 AI 기반 도서/무드 분석은 다른 MSA 컴포넌트가 담당하며 이 계약의 범위 밖이다.
    이 API Server는 이미 인식된 isbn을 넘겨받아 외부 도서 API 조회, 독서 진도 계산,
    서재·스크랩·동물 사서 CRUD 등 단순 데이터베이스 CRUD로 해결 가능한 부분만 다룬다.

    모든 API는 인증이 필요하다. 사용자 식별자는 요청 body가 아니라 검증된 인증 principal에서 얻는다.
  contact:
    name: Virtual Shelf Book Service Team
servers:
  - url: /
    description: 현재 호스트
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/library/books:
    post:
      tags: [Library]
      operationId: createLibraryBook
      summary: 서재 책 등록
      description: |
        사용자가 최종 확인한 도서 정보를 로그인 사용자의 개인 서재에 등록한다.
        알라딘 검색 결과를 그대로 쓰거나, 결과가 없으면(폴백) 사용자가 직접 입력한 값을 그대로 받는다.
        `coverUrl`을 생략하지 않고 검색 결과의 실제 표지 URL을 그대로 전달하면 실제 표지가 기본 표지로 적용된다.
        `shelfId`를 생략하면 사용자의 기본 책장에 등록된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLibraryBookRequest"
            examples:
              전체_입력:
                summary: genre/readingStatus/shelfId까지 모두 지정
                description: |
                  보낸 값이 그대로 저장된다 — 서버 기본값으로 덮어쓰지 않는다.
                  `shelfId`는 로그인 사용자 소유의 책장이어야 한다(아니면 400).
                value:
                  title: 어린 왕자
                  author: 앙투안 드 생텍쥐페리
                  isbn: "9788932917245"
                  genre: LITERARY_FICTION
                  publisher: 열린책들
                  publishedDate: "2015-10-20"
                  totalPages: 160
                  coverUrl: https://example.com/covers/9788932917245.jpg
                  readingStatus: READING
                  shelfId: 1
              선택_필드_생략:
                summary: genre/readingStatus/shelfId 생략 — 서버 기본값 적용
                description: |
                  `genre`는 `NONE`, `readingStatus`는 `PLANNED`로 저장되고
                  사용자의 기본 책장에 등록된다.
                value:
                  title: 어린 왕자
                  author: 앙투안 드 생텍쥐페리
                  isbn: "9788932917245"
                  publisher: 열린책들
                  publishedDate: "2015-10-20"
                  totalPages: 160
                  coverUrl: https://example.com/covers/9788932917245.jpg
      responses:
        "201":
          description: 서재 책 등록 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateLibraryBookResponse"
              example:
                bookId: 123
                shelfId: 1
                shelfRank: "m"
                title: 어린 왕자
                author: 앙투안 드 생텍쥐페리
                genre: NONE
                readingStatus: PLANNED
                currentPage: 0
                progress: 0.0
        "400":
          $ref: "#/components/responses/InvalidCreateBookData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/BookAlreadyRegistered"

    get:
      tags: [Library]
      operationId: getLibraryBooks
      summary: 서재 도서 목록 조회
      description: |
        로그인 사용자의 서재에서 필터, 정렬, 페이지 조건에 맞는 도서를 조회한다.
        `shelfId`를 생략하면 모든 책장을 합산해 조회한다. `shelfId`를 지정하면 `GET /api/v1/library/shelves/{shelfId}/books`와 동일하게 그 책장의 책만 조회한다.
      parameters:
        - name: shelfId
          in: query
          required: false
          description: 특정 책장으로 좁혀 조회. 생략하면 전체 책장 합산
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: author
          in: query
          required: false
          description: 저자명
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: sortBy
          in: query
          required: false
          description: 정렬 기준. SHELF_ORDER는 사용자가 재정렬한 서재 순서(`shelfRank` 오름차순이 자연스러운 진열 순서)다.
          schema:
            type: string
            enum: [SHELF_ORDER, TITLE, AUTHOR, CREATED_AT, PROGRESS]
            default: SHELF_ORDER
        - name: sortOrder
          in: query
          required: false
          description: 정렬 방향
          schema:
            type: string
            enum: [ASC, DESC]
            default: ASC
        - name: page
          in: query
          required: false
          description: 0부터 시작하는 페이지 번호
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: 페이지당 항목 수
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 서재 도서 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBookPageResponse"
              example:
                books:
                  - bookId: 123
                    shelfId: 1
                    shelfRank: "m"
                    title: 살인자의 기억법
                    author: 김영하
                    genre: MYSTERY_THRILLER
                    readingStatus: READING
                    coverUrl: https://example.com/covers/9791198765432.jpg
                    progress: 45.0
                page: 0
                size: 20
                totalElements: 1
                totalPages: 1
        "400":
          $ref: "#/components/responses/InvalidFilterParameter"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/library/books/{bookId}:
    parameters:
      - $ref: "#/components/parameters/BookId"
    get:
      tags: [Library]
      operationId: getLibraryBook
      summary: 서재 책 상세 조회
      description: 로그인 사용자의 서재에서 특정 도서의 상세 정보를 조회한다.
      responses:
        "200":
          description: 서재 책 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryBookDetailResponse"
              example:
                bookId: 123
                shelfId: 1
                shelfRank: "m"
                title: 어린 왕자
                author: 앙투안 드 생텍쥐페리
                isbn: "9788932917245"
                genre: NONE
                publisher: 열린책들
                totalPages: 160
                readingStatus: PLANNED
                coverUrl: https://example.com/covers/9788932917245.jpg
                currentPage: 50
                progress: 31.25
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibraryBookViewAccessDenied"
        "404":
          $ref: "#/components/responses/LibraryBookViewNotFound"

    patch:
      tags: [Library]
      operationId: updateLibraryBook
      summary: 서재 책 정보 수정
      description: |
        로그인 사용자의 서재에 등록된 도서 정보를 수정한다. 9개 필드를 항상 모두 포함해야 하며,
        `isbn`/`publisher`/`publishedDate`/`coverUrl`/`totalPages`에 `null`을 보내면 그 값을 삭제한다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLibraryBookRequest"
            example:
              title: 어린 왕자
              author: 생택쥐페리
              isbn: "9788932917245"
              publisher: 열린책들
              publishedDate: "2015-10-20"
              coverUrl: https://example.com/covers/9788932917245.jpg
              totalPages: 158
      responses:
        "200":
          description: 서재 책 정보 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateLibraryBookResponse"
              example:
                bookId: 123
                shelfRank: "m"
                title: 어린 왕자
                genre: NONE
                readingStatus: PLANNED
                totalPages: 158
                updatedAt: "2026-08-17T01:10:00+09:00"
        "400":
          $ref: "#/components/responses/InvalidUpdateBookData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibraryBookUpdateAccessDenied"
        "404":
          $ref: "#/components/responses/LibraryBookUpdateNotFound"

    delete:
      tags: [Library]
      operationId: deleteLibraryBook
      summary: 서재 책 삭제
      description: 로그인 사용자의 개인 서재에서 지정한 책을 삭제한다.
      responses:
        "204":
          description: 서재 책 삭제 성공. 응답 body 없음
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibraryBookDeleteAccessDenied"
        "404":
          $ref: "#/components/responses/LibraryBookDeleteNotFound"
  /api/v1/library/books/{bookId}/order:
    parameters:
      - $ref: "#/components/parameters/BookId"
    patch:
      tags: [Library]
      operationId: reorderLibraryBook
      summary: 서재 책 순서 변경
      description: |
        지정한 책을 같은 책장 내 다른 책(`beforeBookId` 또는 `afterBookId`)의 앞/뒤로 옮긴다.
        서버가 두 이웃 책 사이에 들어갈 새 `shelfRank`를 계산해 저장하므로, 클라이언트는 `shelfRank` 값을 직접 계산하거나 지정하지 않는다.
        맨 앞/맨 뒤로 옮기려면 클라이언트가 이미 조회해 알고 있는 현재 첫 번째/마지막 책을 기준으로 지정한다.
        `shelfRank`는 책장별로 유일하므로, 다른 책장에 속한 책을 기준으로 지정할 수 없다(다른 책장으로 옮기려면 `moveLibraryBookToShelf`를 사용한다).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderLibraryBookRequest"
            example:
              afterBookId: 456
      responses:
        "200":
          description: 서재 책 순서 변경 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReorderLibraryBookResponse"
              example:
                bookId: 123
                shelfRank: "mV"
                updatedAt: "2026-08-20T01:10:00+09:00"
        "400":
          $ref: "#/components/responses/InvalidReorderTarget"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibraryBookReorderAccessDenied"
        "404":
          $ref: "#/components/responses/LibraryBookReorderNotFound"
  /api/v1/library/books/{bookId}/shelf:
    parameters:
      - $ref: "#/components/parameters/BookId"
    patch:
      tags: [Shelf]
      operationId: moveLibraryBookToShelf
      summary: 책장에 책 넣기 (책 이동)
      description: |
        지정한 책을 다른 책장(`shelfId`)으로 옮긴다. 이동한 책은 대상 책장의 맨 뒤(`ShelfRank.after`)에 배치된다.
        대상 책장은 이 책을 소유한 사용자의 책장이어야 하며, 존재하지 않거나 다른 사용자 소유면 400(`INVALID_SHELF_TARGET`)이다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveLibraryBookToShelfRequest"
            example:
              shelfId: 2
      responses:
        "200":
          description: 책장 이동 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveLibraryBookToShelfResponse"
              example:
                bookId: 123
                shelfId: 2
                shelfRank: "m"
                updatedAt: "2026-08-21T09:00:00+09:00"
        "400":
          $ref: "#/components/responses/InvalidShelfTarget"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibraryBookMoveAccessDenied"
        "404":
          $ref: "#/components/responses/LibraryBookMoveNotFound"
  /api/v1/library/books/{bookId}/progress:
    parameters:
      - $ref: "#/components/parameters/BookId"
    patch:
      tags: [Reading Progress]
      operationId: updateReadingProgress
      summary: 독서 진도 수정
      description: |
        현재/전체 페이지를 수정하고 진도율을 서버에서 계산한다. `totalPages`를 모르면 생략할 수 있고,
        이 경우 `progress`는 null로 응답된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReadingProgressRequest"
            example:
              currentPage: 80
              totalPages: 160
      responses:
        "200":
          description: 독서 진도 수정 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateReadingProgressResponse"
              example:
                bookId: 123
                currentPage: 80
                totalPages: 160
                progress: 50.0
                updatedAt: "2026-08-17T21:30:00+09:00"
        "400":
          $ref: "#/components/responses/InvalidPageValue"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/LibraryBookProgressAccessDenied"
        "404":
          $ref: "#/components/responses/LibraryBookProgressNotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 인증 서비스가 발급하고 Book Service가 검증하는 access token
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      description: 개인 서재 도서 식별자
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 123
...
  schemas:
...
    Genre:
      type: string
      enum:
        [
          NONE,
          SCIENCE_FICTION,
          FANTASY,
          ROMANCE,
          MYSTERY_THRILLER,
          LITERARY_FICTION,
          ESSAY,
          POETRY_DRAMA,
          HUMANITIES,
          HISTORY,
          BUSINESS_ECONOMICS,
          SELF_HELP,
          SCIENCE,
          ARTS,
          RELIGION,
          COMPUTER_IT,
        ]
      description: 도서 장르. 생략하면 NONE으로 저장된다.
      default: NONE
    ReadingStatus:
      type: string
      enum: [PLANNED, READING, COMPLETED]
      description: 독서 상태. 생략하면 PLANNED로 저장된다. 진도율(progress)과 자동으로 연동되지 않는 별도 필드다.
      default: PLANNED
    CreateLibraryBookRequest:
      type: object
      required: [title, author]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        author:
          type: string
          minLength: 1
          maxLength: 100
        isbn:
          type: string
          pattern: "^(?:[0-9]{10}|[0-9]{13})$"
        genre:
          $ref: "#/components/schemas/Genre"
        publisher:
          type: string
          maxLength: 100
        publishedDate:
          type: string
          format: date
        totalPages:
          type: integer
          format: int32
          minimum: 1
          description: 생략 가능. 알라딘 API가 제공하지 않는 도서가 많아 등록 시점에 모를 수 있다.
        coverUrl:
          type: string
          format: uri
        readingStatus:
          $ref: "#/components/schemas/ReadingStatus"
        shelfId:
          type: integer
          format: int64
          minimum: 1
          description: 등록할 책장. 생략하면 사용자의 기본 책장에 등록한다.
...
    CreateLibraryBookResponse:
      type: object
      additionalProperties: false
      required: [bookId, shelfId, shelfRank, title, author, genre, readingStatus, currentPage, progress]
      properties:
        bookId:
          type: integer
          format: int64
          minimum: 1
        shelfId:
          type: integer
          format: int64
          minimum: 1
        shelfRank:
          $ref: "#/components/schemas/ShelfRank"
        title:
          type: string
        author:
          type: string
        genre:
          $ref: "#/components/schemas/Genre"
        readingStatus:
          $ref: "#/components/schemas/ReadingStatus"
        currentPage:
          type: integer
          format: int32
          minimum: 0
        progress:
          type: [number, "null"]
          format: double
          minimum: 0
          maximum: 100
          description: totalPages를 모르면 null이다.
...
    LibraryBookPageResponse:
      type: object
      additionalProperties: false
      required: [books, page, size, totalElements, totalPages]
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/LibraryBookSummary"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
    LibraryBookDetailResponse:
      type: object
      additionalProperties: false
      required: [bookId, shelfId, shelfRank, title, author, isbn, genre, publisher, publishedDate, totalPages, coverUrl, readingStatus, currentPage, progress]
      properties:
        bookId:
          type: integer
          format: int64
          minimum: 1
        shelfId:
          type: integer
          format: int64
          minimum: 1
        shelfRank:
          $ref: "#/components/schemas/ShelfRank"
        title:
          type: string
        author:
          type: string
        isbn:
          type: [string, "null"]
          pattern: "^(?:[0-9]{10}|[0-9]{13})$"
        genre:
          $ref: "#/components/schemas/Genre"
        publisher:
          type: [string, "null"]
        publishedDate:
          type: [string, "null"]
          format: date
        totalPages:
          type: [integer, "null"]
          format: int32
          minimum: 1
        coverUrl:
          type: [string, "null"]
          format: uri
        readingStatus:
          $ref: "#/components/schemas/ReadingStatus"
        currentPage:
          type: integer
          format: int32
          minimum: 0
        progress:
          type: [number, "null"]
          format: double
          minimum: 0
          maximum: 100
          description: totalPages를 모르면 null이다.
    UpdateLibraryBookRequest:
      type: object
      required: [title, author, isbn, genre, publisher, publishedDate, coverUrl, readingStatus, totalPages]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        author:
          type: string
          minLength: 1
          maxLength: 100
        isbn:
          type: [string, "null"]
          pattern: "^(?:[0-9]{10}|[0-9]{13})$"
        genre:
          $ref: "#/components/schemas/Genre"
        publisher:
          type: [string, "null"]
          minLength: 1
          maxLength: 100
        publishedDate:
          type: [string, "null"]
          format: date
        coverUrl:
          type: [string, "null"]
          format: uri
        readingStatus:
          $ref: "#/components/schemas/ReadingStatus"
        totalPages:
          type: [integer, "null"]
          format: int32
          minimum: 1
      description: |
        9개 필드를 항상 모두 포함해야 한다(부분 생략 불가). `isbn`/`publisher`/`publishedDate`/`coverUrl`/`totalPages`는
        `null`이면 그 값을 삭제하고, 값이 있으면 교체한다. `title`/`author`/`genre`/`readingStatus`는 `null`을 허용하지 않는다.
    UpdateLibraryBookResponse:
      type: object
      additionalProperties: false
      required: [bookId, shelfRank, title, genre, coverUrl, readingStatus, totalPages, updatedAt]
      properties:
        bookId:
          type: integer
          format: int64
          minimum: 1
        shelfRank:
          $ref: "#/components/schemas/ShelfRank"
        title:
          type: string
        genre:
          $ref: "#/components/schemas/Genre"
        coverUrl:
          type: [string, "null"]
          format: uri
        readingStatus:
          $ref: "#/components/schemas/ReadingStatus"
        totalPages:
          type: [integer, "null"]
          format: int32
          minimum: 1
        updatedAt:
          type: string
          format: date-time
    ReorderLibraryBookRequest:
      type: object
      minProperties: 1
      maxProperties: 1
      properties:
        beforeBookId:
          type: integer
          format: int64
          minimum: 1
        afterBookId:
          type: integer
          format: int64
          minimum: 1
      description: |
        `beforeBookId` 또는 `afterBookId` 중 정확히 하나만 지정한다. 대상 책을 그 책의 바로 앞/뒤로 옮긴다.
        지정한 책은 같은 사용자 서재에 존재해야 하며, 대상 책 자신을 지정할 수 없다.
    ReorderLibraryBookResponse:
      type: object
      additionalProperties: false
      required: [bookId, shelfRank, updatedAt]
      properties:
        bookId:
          type: integer
          format: int64
          minimum: 1
        shelfRank:
          $ref: "#/components/schemas/ShelfRank"
        updatedAt:
          type: string
          format: date-time
    MoveLibraryBookToShelfRequest:
      type: object
      required: [shelfId]
      properties:
        shelfId:
          type: integer
          format: int64
          minimum: 1
      description: 이동할 대상 책장. 이 사용자의 책장이어야 한다.
    MoveLibraryBookToShelfResponse:
      type: object
      additionalProperties: false
      required: [bookId, shelfId, shelfRank, updatedAt]
      properties:
        bookId:
          type: integer
          format: int64
          minimum: 1
        shelfId:
          type: integer
          format: int64
          minimum: 1
        shelfRank:
          $ref: "#/components/schemas/ShelfRank"
        updatedAt:
          type: string
          format: date-time
...
    UpdateReadingProgressRequest:
      type: object
      required: [currentPage]
      properties:
        currentPage:
          type: integer
          format: int32
          minimum: 0
        totalPages:
          type: [integer, "null"]
          format: int32
          minimum: 1
          description: 생략(null)하면 기존에 등록된 totalPages를 지운다. currentPage는 totalPages보다 클 수 없다.
      description: totalPages를 생략하면 currentPage 범위 검증(0 이상)만 적용된다.
    UpdateReadingProgressResponse:
      type: object
      additionalProperties: false
      required: [bookId, currentPage, totalPages, progress, updatedAt]
      properties:
        bookId:
          type: integer
          format: int64
          minimum: 1
        currentPage:
          type: integer
          format: int32
          minimum: 0
        totalPages:
          type: [integer, "null"]
          format: int32
          minimum: 1
        progress:
          type: [number, "null"]
          format: double
          minimum: 0
          maximum: 100
          description: totalPages를 모르면 null이다.
        updatedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: UNAUTHORIZED
            message: 인증이 필요합니다.
...
    InvalidCreateBookData:
      description: 서재 등록 필수 도서 정보 누락
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INVALID_BOOK_DATA
            message: 필수 도서 정보가 누락되었습니다.
    BookAlreadyRegistered:
      description: 사용자 서재에 이미 등록된 도서
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: BOOK_ALREADY_REGISTERED
            message: 이미 서재에 등록된 도서입니다.
    InvalidFilterParameter:
      description: 올바르지 않은 목록 필터 또는 정렬 조건
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INVALID_FILTER_PARAMETER
            message: 올바르지 않은 필터 또는 정렬 조건입니다.
    LibraryBookViewAccessDenied:
      description: 서재 도서 조회 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_ACCESS_DENIED
            message: 해당 도서에 접근할 권한이 없습니다.
    LibraryBookViewNotFound:
      description: 조회할 서재 도서 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_NOT_FOUND
            message: 서재에서 해당 도서를 찾을 수 없습니다.
    InvalidUpdateBookData:
      description: 올바르지 않은 도서 수정 정보
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INVALID_BOOK_DATA
            message: 올바르지 않은 도서 정보입니다.
    LibraryBookUpdateAccessDenied:
      description: 서재 도서 수정 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_ACCESS_DENIED
            message: 해당 도서를 수정할 권한이 없습니다.
    LibraryBookUpdateNotFound:
      description: 수정할 서재 도서 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_NOT_FOUND
            message: 수정할 도서를 찾을 수 없습니다.
    LibraryBookDeleteAccessDenied:
      description: 서재 도서 삭제 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_ACCESS_DENIED
            message: 해당 도서를 삭제할 권한이 없습니다.
    LibraryBookDeleteNotFound:
      description: 삭제할 서재 도서 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_NOT_FOUND
            message: 삭제할 도서를 찾을 수 없습니다.
    InvalidReorderTarget:
      description: 기준 책(beforeBookId/afterBookId)이 자기 자신이거나 같은 책장에 속하지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INVALID_REORDER_TARGET
            message: 기준으로 지정한 책을 찾을 수 없거나 같은 책장에 속하지 않습니다.
    LibraryBookReorderAccessDenied:
      description: 서재 도서 순서 변경 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_ACCESS_DENIED
            message: 해당 도서의 순서를 변경할 권한이 없습니다.
    LibraryBookReorderNotFound:
      description: 순서를 변경할 서재 도서 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_NOT_FOUND
            message: 순서를 변경할 도서를 찾을 수 없습니다.
    InvalidPageValue:
      description: 현재/전체 페이지 값이 도메인 규칙을 위반함
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INVALID_PAGE_VALUE
            message: 현재 페이지와 전체 페이지 값을 확인해야 합니다.
    LibraryBookProgressAccessDenied:
      description: 독서 진도 수정 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_ACCESS_DENIED
            message: 해당 도서의 독서 정보를 수정할 권한이 없습니다.
    LibraryBookProgressNotFound:
      description: 진도를 수정할 서재 도서 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_NOT_FOUND
            message: 서재에서 해당 도서를 찾을 수 없습니다.
    InvalidShelfTarget:
      description: 대상 책장(shelfId)이 존재하지 않거나 이 사용자의 책장이 아님
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INVALID_SHELF_TARGET
            message: 대상 책장을 찾을 수 없거나 이 사용자의 책장이 아닙니다.
    LibraryBookMoveAccessDenied:
      description: 책장 이동 권한 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_ACCESS_DENIED
            message: 해당 도서의 책장을 변경할 권한이 없습니다.
    LibraryBookMoveNotFound:
      description: 책장을 이동할 서재 도서 없음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: LIBRARY_BOOK_NOT_FOUND
            message: 책장을 이동할 도서를 찾을 수 없습니다.
...